requirements.txt
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-cors-headers==4.4.0
pymongo==4.8.0
python-dotenv==1.0.1
openpyxl==3.1.5
gunicorn==22.0.0
